requirements.txt
langchain_openai
langchain
langchain-community
openai
azure-cognitiveservices-speech
streamlit
wxPython
pillow
